requirements.txt
google-adk>=0.1.0
google-genai>=0.1.0
opentelemetry-instrumentation-google-genai>=0.1.0
python-dotenv>=1.0.0
kaggle>=1.5.0